refactor(frontend): move util function calculateTokenUsdBalance
