+++ src/frontend/src/tests/lib/utils/tokens.utils.spec.ts
@@ -7,7 +7,6 @@ import type { ExchangesData } from '$lib/types/exchange';
 import type { Token, TokenToPin, TokenUi } from '$lib/types/token';
 import { usdValue } from '$lib/utils/exchange.utils';
 import {
-	calculateTokenUsdBalance,
 	filterEnabledTokens,
 	pinTokensWithBalanceAtTop,
 	sortTokens,
@@ -321,33 +320,6 @@ describe('filterEnabledTokens', () => {
 
 		const result = filterEnabledTokens([tokens]);
 		expect(result).toEqual([ENABLED_BY_DEFAULT_ICP_TOKEN, ENABLED_BY_DEFAULT_ETHEREUM_TOKEN]);
-	});
-});
-
-describe('calculateTokenUsdBalance', () => {
-	const mockUsdValue = usdValue as MockedFunction<typeof usdValue>;
-
-	beforeEach(() => {
-		vi.resetAllMocks();
-
-		mockUsdValue.mockImplementation(
-			({ balance, exchangeRate }) => Number(balance ?? 0) * exchangeRate
-		);
-	});
-
-	it('should correctly calculate USD balance for the token', () => {
-		const result = calculateTokenUsdBalance({ token: ETHEREUM_TOKEN, $balances, $exchanges });
-		expect(result).toEqual(bn3.toNumber());
-	});
-
-	it('should return undefined if exchange rate is not available', () => {
-		const result = calculateTokenUsdBalance({ token: ICP_TOKEN, $balances, $exchanges: {} });
-		expect(result).toEqual(undefined);
-	});
-
-	it('should return 0 if balances store is not available', () => {
-		const result = calculateTokenUsdBalance({ token: ETHEREUM_TOKEN, $balances: {}, $exchanges });
-		expect(result).toEqual(0);
 	});
 });
